# perf: initialization and builins creation