--- a/index.js
+++ b/index.js
@@ -37,39 +37,38 @@ const permanentModules = [
   'zlib'
 ]
 
-const versionLockedModules = {
-  freelist: '<6.0.0',
-  v8: '>=1.0.0',
-  process: '>=1.1.0',
-  inspector: '>=8.0.0',
-  async_hooks: '>=8.1.0',
-  http2: '>=8.4.0',
-  perf_hooks: '>=8.5.0',
-  trace_events: '>=10.0.0',
-  worker_threads: '>=12.0.0',
-  'node:test': '>=18.0.0'
-}
+const versionLockedModules = [
+  'node:test', 'worker_threads', 'trace_events', 'perf_hooks', 'http2', 'async_hooks', 'inspector', 'process', 'v8'
+]
+const versionLockedModulesRange = [
+  '>=18.0.0', '>=12.0.0', '>=10.0.0', '>=8.5.0', '>=8.4.0', '>=8.1.0', '>=8.0.0', '>=1.1.0', '>=1.0.0'
+]
 
-const experimentalModules = {
-  worker_threads: '>=10.5.0',
-  wasi: '>=12.16.0',
-  diagnostics_channel: '^14.17.0 || >=15.1.0'
-}
+const experimentalModulesRanges = [
+  '^14.17.0 || >=15.1.0', '>=12.16.0', '>=10.5.0'
+]
+const experimentalModules = ['diagnostics_channel', 'wasi', 'worker_threads']
 
 module.exports = ({ version = process.version, experimental = false } = {}) => {
-  const builtins = [...permanentModules]
+  const builtins = permanentModules.slice()
 
-  for (const [name, semverRange] of Object.entries(versionLockedModules)) {
-    if (version === '*' || satisfies(version, semverRange)) {
-      builtins.push(name)
+  for (let i = 0; i < versionLockedModulesRange.length; i++) {
+    if (version === '*' || satisfies(version, versionLockedModulesRange[i])) {
+      Array.prototype.push.apply(builtins, versionLockedModules.slice(i))
+      break
     }
   }
 
+  if (version === '*' || satisfies(version, '<6.0.0')) {
+    builtins.push('freelist')
+  }
+
   if (experimental) {
-    for (const [name, semverRange] of Object.entries(experimentalModules)) {
+    for (let i = 0; i < experimentalModulesRanges.length; i++) {
+      const name = experimentalModules[i]
       if (
         !builtins.includes(name) &&
-        (version === '*' || satisfies(version, semverRange))
+        (version === '*' || satisfies(version, experimentalModulesRanges[i]))
       ) {
         builtins.push(name)
       }