@juliangruber 2024-04-07
Could you please explain why this optimization is important to you?