@H4ad 2024-04-07
I'm exploring the performance opportunities of `npm run`, and this module is required by one of the dependencies of NPM and it takes `5ms` to load because it triggers the load of the entire `semver`:

![image](https://github.com/juliangruber/builtins/assets/12551007/6bbced47-55e4-4fd1-aeb2-5953fbd6c52e)