@juliangruber 2024-04-07
Got it! That's important indeed. What do you think about only changing the `semver` require, and keeping the rest of the logic simple as is? I assume this will yield basically the same benefit, while keeping the code base simple.