@H4ad 2024-04-07
I created another PR just for `semver` #53, but why not pushing also the optimization on the built-ins creation? Since you have a good amount of tests, this new approach is not that complicated and doubles the performance for most cases (even though it's already fast)