@juliangruber 2024-04-07
Yes it makes it faster, but is it relevant? Otoh, with the data structures from the refactor the code becomes harder to maintain, so it's not worth it to me